\name{forecast.grid}
\docType{data}
\alias{forecast.grid}
\title{48-hour Temperature Forecast for the North American Pacific Northwest}

\description{
This data set gives the 48-hour Temperature Forecast verifying January 12, 
2002 at 0 hours GMT for the North American Pacific Northwest as provided 
by the Aviation (AVN) member of the University of Washington MM5 Mesoscale 
Ensemble, using a 12-km grid resolution.
}

\format{
\emph{gridded forecast} a numeric vector with 10098 forecasted temperature.
}

\source{http://metoc.apl.washington.edu/uwme/}

\references{
Gel, Y., Raftery, A. E., Gneiting, T. (2004). Calibrated 
mesoscale weather field forecasting: The Geostatistical Output 
Perturbation (GOP) method (with discussion). \emph{Journal of the American 
Statistical Association}, \bold{Vol. 99 (467)}, 575--583.

Grimit, E. P., Mass, C. F. (2002). Initial results of a short-range 
ensemble forecasting system over the Pacific Northwest. 
\emph{Weather and Forecasting} \bold{17}, 192--205.
}

\keyword{datasets}
