\name{gridlat}
\docType{data}
\alias{gridlat}
\title{Latitudes of localities in the North American Pacific Northwest}

\description{
This data set gives the latitudes of the grid points for the 12-km domain 
of the University of Washington MM5 Mesoscale Ensemble.
}

\format{
\emph{gridded.lat} a numeric vector with 10098 latitudes.
}

\source{http://metoc.apl.washington.edu/uwme/}

\references{
Gel, Y., Raftery, A. E., Gneiting, T. (2004). Calibrated 
mesoscale weather field forecasting: The Geostatistical Output 
Perturbation (GOP) method (with discussion). \emph{Journal of the American 
Statistical Association}, \bold{Vol. 99 (467)}, 575--583.

Grimit, E. P., Mass, C. F. (2002). Initial results of a mesoscale 
short-range ensemble forecasting system over the Pacific Northwest. 
\emph{Weather and Forecasting} \bold{17}, 192--205.
} 

\keyword{datasets}
